requirements.txt
nonebot2~=2.3.2
nonebot-adapter-onebot~=2.4.4
yaml~=0.2.5
pyyaml~=6.0.1
gitpython~=3.1.43
pydantic~=2.8.2